requirements.txt
torch
diffusers
pillow
requests
transformers
accelerate